// src/enum/dbConfig.enum.ts
export enum DbConfig {
    TYPE = 'DB_TYPE',
    HOST = 'DB_HOST',
    PORT = 'DB_PORT',
    USERNAME = 'DB_USERNAME',
    PASSWORD = 'DB_PASSWORD',
    DATABASE = 'DB_DATABASE',
    RETRYDELAY = 'DB_RETRYDELAY',
    RETRYATTEMPTS = 'DB_RETRYATTEMPTS',
    SYNCHRONIZE = 'DB_SYNCHRONIZE',
    AUTOLOADENTITIES = 'DB_AUTOLOADENTITIES',
}